Remove old files; add LGR; implement outfits
