+++ src/outfit/index.ts
@@ -1,0 +1,11 @@
+export { usingThumbRing } from "./bonusgear";
+export { meatOutfit } from "./meat";
+export { freeFightOutfit } from "./free";
+export {
+  usingPurse,
+  refreshLatte,
+  useUPCs,
+  tryFillLatte,
+  waterBreathingEquipment,
+  familiarWaterBreathingEquipment,
+} from "./lib";
